환경에 따른 인증번호 생성 로직 수정
